requirements.txt
fastapi
uvicorn
google-generativeai
numpy
pydantic
